chore: create insight-handler class and move logic
